contour contains point with tests added

+++ include/cg/operations/convex.h
@@ -19,7 +19,7 @@ namespace cg
       v2++;
       contour_2::circulator_t v3 = v2;
       v3++;
-      for (int i = 0; i < c.vertices_num(); i++, v1++, v2++, v3++)
+      for (size_t i = 0; i < c.vertices_num(); i++, v1++, v2++, v3++)
       {
          if (orientation(*v1, *v2, *v3) == CG_RIGHT)
          {
